merge v1.9.8 to 2.0-dev-sharding

+++ src/main/java/org/teasoft/honey/osql/core/ObjSQLRich.java
@@ -7,15 +7,23 @@
 package org.teasoft.honey.osql.core;
 
 import java.util.List;
+import java.util.Map;
 
 import org.teasoft.bee.osql.Condition;
 import org.teasoft.bee.osql.FunctionType;
 import org.teasoft.bee.osql.IncludeType;
+import org.teasoft.bee.osql.MapSql;
+import org.teasoft.bee.osql.MapSqlKey;
+import org.teasoft.bee.osql.MapSqlSetting;
+import org.teasoft.bee.osql.MapSuid;
 import org.teasoft.bee.osql.ObjSQLException;
 import org.teasoft.bee.osql.ObjToSQLRich;
 import org.teasoft.bee.osql.OrderType;
 import org.teasoft.bee.osql.SuidRich;
+import org.teasoft.bee.osql.exception.BeeErrorGrammarException;
 import org.teasoft.bee.osql.exception.BeeIllegalParameterException;
+import org.teasoft.honey.osql.name.NameUtil;
+import org.teasoft.honey.util.SuidHelper;
 
 /**
  * @author Kingstar
@@ -23,9 +31,11 @@ import org.teasoft.bee.osql.exception.BeeIllegalParameterException;
  */
 public class ObjSQLRich extends ObjSQL implements SuidRich {
 
-	private ObjToSQLRich objToSQLRich; // = BeeFactory.getHoneyFactory().getObjToSQLRich();
-//	private BeeSql beeSql = BeeFactory.getHoneyFactory().getBeeSql();
+    private ObjToSQLRich objToSQLRich; 
 	
+    public ObjSQLRich() {
+    }
+    
 	public ObjToSQLRich getObjToSQLRich() {
 		if(objToSQLRich==null) objToSQLRich=BeeFactory.getHoneyFactory().getObjToSQLRich();
 		return objToSQLRich;
@@ -112,43 +122,51 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 	@Override
 	public <T> int insert(T entity[]) {
 		if (entity == null || entity.length<1) return -1;
+		checkNull(entity);
 //		int len = entity.length;
 //		String insertSql[] = new String[len];
 		String insertSql[] = getObjToSQLRich().toInsertSQL(entity);
-		_regEntityClass(entity[0]);
+		_regEntityClass1(entity[0]);
+		
+		HoneyUtil.revertId(entity);
 		
 		return getBeeSql().batch(insertSql);
+	}
+	
+	private <T> void checkNull(T entity[]) {
+		for (int i = 0; i < entity.length; i++) {
+			if(entity[i]==null) throw new ObjSQLException("entity[] have null element, index: "+i);
+		}
 	}
 
 	@Override
 	public <T> int insert(T entity[], String excludeFields) {
 		if (entity == null || entity.length<1) return -1;
+		checkNull(entity);
 //		int len = entity.length;
 //		String insertSql[] = new String[len];
 		String insertSql[] = getObjToSQLRich().toInsertSQL(entity, excludeFields);
-		_regEntityClass(entity[0]);
+		_regEntityClass1(entity[0]);
+		
+		HoneyUtil.revertId(entity);
 		
 		return getBeeSql().batch(insertSql);
 	}
 
 	@Override
 	public <T> int insert(T entity[], int batchSize) {
-		if (entity == null || entity.length<1) return -1;
-//		int len = entity.length;
-//		String insertSql[] = new String[len];
-		String insertSql[] = getObjToSQLRich().toInsertSQL(entity,batchSize);
-		_regEntityClass(entity[0]);
-
-		return getBeeSql().batch(insertSql, batchSize);
+		return insert(entity, batchSize, "");
 	}
 
 	@Override
 	public <T> int insert(T entity[], int batchSize, String excludeFields) {
 		if (entity == null || entity.length<1) return -1;
-//		int len = entity.length;
-//		String insertSql[] = new String[len];
+		checkNull(entity);
+		if(batchSize<=0) batchSize=10;
 		String insertSql[] = getObjToSQLRich().toInsertSQL(entity,batchSize, excludeFields);
-		_regEntityClass(entity[0]);
+		_regEntityClass1(entity[0]);
+		
+		HoneyUtil.revertId(entity);
 		
 		return getBeeSql().batch(insertSql, batchSize);
 	}
@@ -159,7 +177,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		int r = 0;
 		String sql = getObjToSQLRich().toUpdateSQL(entity, updateFields);
 		Logger.logSQL("update SQL(updateFields) :", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		r = getBeeSql().modify(sql);
 
 		return r;
@@ -176,17 +194,28 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 
 	@Override
 	public <T> String selectWithFun(T entity, FunctionType functionType, String fieldForFun) {
+		return selectWithFun(entity, functionType, fieldForFun, null);
+	}
+
+	@Override
+	public <T> String selectWithFun(T entity, FunctionType functionType, String fieldForFun, Condition condition) {
 		if (entity == null) return null;
 		String s = null;
-		try {
-			String sql = getObjToSQLRich().toSelectFunSQL(entity, functionType, fieldForFun);
-			_regEntityClass(entity);
-			s=getBeeSql().selectFun(sql);
-		} catch (ObjSQLException e) {
-			throw e;
-		}
-
+		String sql = getObjToSQLRich().toSelectFunSQL(entity, functionType, fieldForFun, condition);
+		_regEntityClass1(entity);
+		s = getBeeSql().selectFun(sql);
 		return s;
+	}
+
+	@Override
+	public <T> int count(T entity) {
+		return count(entity, null);
+	}
+
+	@Override
+	public <T> int count(T entity, Condition condition) {
+		String total = selectWithFun(entity, FunctionType.COUNT, "*", condition);
+		return total == null ? 0 : Integer.parseInt(total);
 	}
 
 	@Override
@@ -195,7 +224,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		int r = 0;
 		String sql = getObjToSQLRich().toUpdateSQL(entity, updateFields, includeType);
 		Logger.logSQL("update SQL(updateFields) :", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		r = getBeeSql().modify(sql);
 
 		return r;
@@ -214,7 +243,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		if (entity == null) return -1;
 		String sql = getObjToSQLRich().toUpdateSQL(entity, includeType);
 		Logger.logSQL("update SQL: ", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		return getBeeSql().modify(sql);
 	}
 
@@ -223,7 +252,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		if (entity == null) return -1;
 		String sql = getObjToSQLRich().toInsertSQL(entity, includeType);
 		Logger.logSQL("insert SQL: ", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		return getBeeSql().modify(sql);
 	}
 
@@ -232,7 +261,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		if (entity == null) return -1;
 		String sql = getObjToSQLRich().toDeleteSQL(entity, includeType);
 		Logger.logSQL("delete SQL: ", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		return getBeeSql().modify(sql);
 	}
 
@@ -244,7 +273,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		List<String[]> list = null;
 		String sql = getObjToSQLRich().toSelectSQL(entity);
 		Logger.logSQL("List<String[]> select SQL: ", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		list = getBeeSql().select(sql);
 
 		return list;
@@ -258,7 +287,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		List<String[]> list = null;
 		try {
 			String sql = getObjToSQLRich().toSelectSQL(entity, selectFields);
-			_regEntityClass(entity);
+			_regEntityClass1(entity);
 			list = getBeeSql().select(sql);
 		} catch (ObjSQLException e) {
 			throw e;
@@ -266,7 +295,16 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 
 		return list;
 	}
-
+	
+	@Override
+	public <T> List<String[]> selectString(T entity, Condition condition) {
+		if (entity == null) return null;
+		String sql = getObjToSQLRich().toSelectSQL(entity, condition.getIncludeType(), condition);
+		Logger.logSQL("select SQL: ", sql);
+		_regEntityClass1(entity);
+		return getBeeSql().select(sql);
+	}
+	
 	@Override
 	public <T> String selectJson(T entity) {
 		
@@ -274,7 +312,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		
 		String sql = getObjToSQLRich().toSelectSQL(entity);
 		Logger.logSQL("selectJson SQL: ", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		return getBeeSql().selectJson(sql);
 	}
 
@@ -284,38 +322,109 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		
 		String sql = getObjToSQLRich().toSelectSQL(entity, includeType);
 		Logger.logSQL("selectJson SQL: ", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
+		return getBeeSql().selectJson(sql);
+	}
+	
+	@Override
+	public <T> String selectJson(T entity, String selectField) {
+		if (entity == null) return null;
+		
+		String sql = getObjToSQLRich().toSelectSQL(entity, selectField);
+		Logger.logSQL("selectJson(T entity, String selectField) SQL: ", sql);
+		_regEntityClass1(entity);
+		return getBeeSql().selectJson(sql);
+	}
+	
+	@Override
+	public <T> String selectJson(T entity, String selectFields, int start, int size) {
+		if (entity == null) return null;
+		if(size<=0) throw new BeeIllegalParameterException("Parameter 'size' need great than 0!");
+		if(start<0) throw new BeeIllegalParameterException("Parameter 'start' need great equal 0!");
+		
+		String sql = getObjToSQLRich().toSelectSQL(entity, selectFields,start,size);
+		
+		Logger.logSQL("selectJson(T entity, String selectField, int start, int size) SQL: ", sql);
+		_regEntityClass1(entity);
 		return getBeeSql().selectJson(sql);
 	}
 
 	@Override
-	public <T> List<T> selectById(T entity, Integer id) {
+	public <T> T selectById(T entity, Integer id) {
 		if (entity == null) return null;
+
+		if (id==null) {
+			Logger.warn("in method selectById,id is null! ");
+			return null;
+		}
+		
 		String sql = getObjToSQLRich().toSelectByIdSQL(entity, id);
 		Logger.logSQL("selectById SQL: ", sql);
-		return getBeeSql().select(sql, entity);
+		List<T> list = getBeeSql().select(sql, entity);
+		
+		return getIdEntity(list);
 	}
-
+	
 	@Override
-	public <T> List<T> selectById(T entity, Long id) {
+	public <T> T selectById(T entity, Long id) {
 		if (entity == null) return null;
+
+		if (id==null) {
+			Logger.warn("in method selectById,id is null! ");
+			return null;
+		}
 		String sql = getObjToSQLRich().toSelectByIdSQL(entity, id);
 		Logger.logSQL("selectById SQL: ", sql);
-		return getBeeSql().select(sql, entity);
+		List<T> list = getBeeSql().select(sql, entity);
+		
+		return getIdEntity(list);
 	}
-
+	
 	@Override
-	public <T> List<T> selectById(T entity, String ids) {
+	public <T> T selectById(T entity, String id) {
 		if (entity == null) return null;
+
+		if (id==null) {
+			Logger.warn("in method selectById,id is null! ");
+			return null;
+		}
+		if(id.contains(",")) {
+			throw new BeeIllegalParameterException("The parameter 'id' of method selectById does not allow to contain comma!");
+		}
+		String sql = getObjToSQLRich().toSelectByIdSQL(entity, id);
+		Logger.logSQL("selectById SQL: ", sql);
+		
+		List<T> list = getBeeSql().select(sql, entity);
+		
+		return getIdEntity(list);
+	}
+	
+	@Override
+	public <T> List<T> selectByIds(T entity, String ids) {
+		if (entity == null) return null;
+
+		if (ids==null) {
+			Logger.warn("in method selectByIds,ids is null! ");
+			return null;
+		}
 		String sql = getObjToSQLRich().toSelectByIdSQL(entity, ids);
-		Logger.logSQL("selectById SQL: ", sql);
+		Logger.logSQL("selectByIds SQL: ", sql);
 		return getBeeSql().select(sql, entity);
+	}
+	
+	private <T> T getIdEntity(List<T> list) {
+		if(list==null || list.size()<1) {
+			return null;
+		}else {
+			return list.get(0);
+		}
 	}
 
 	@Override
 	@SuppressWarnings("rawtypes")
 	public int deleteById(Class c, Integer id) {
-		if (c == null) return 0;
+		if(id==null) Logger.warn("in method deleteById,id is null! ");
+		if (c == null || id==null) return 0;
 		String sql = getObjToSQLRich().toDeleteByIdSQL(c, id);
 		Logger.logSQL("deleteById SQL: ", sql);
 		_regEntityClass2(c);
@@ -325,7 +434,8 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 	@Override
 	@SuppressWarnings("rawtypes")
 	public int deleteById(Class c, Long id) {
-		if (c == null) return 0;
+		if(id==null) Logger.warn("in method deleteById,id is null! ");
+		if (c == null || id==null) return 0;
 		String sql = getObjToSQLRich().toDeleteByIdSQL(c, id);
 		Logger.logSQL("deleteById SQL: ", sql);
 		_regEntityClass2(c);
@@ -335,7 +445,8 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 	@Override
 	@SuppressWarnings("rawtypes")
 	public int deleteById(Class c, String ids) {
-		if (c == null) return 0;
+		if(ids==null) Logger.warn("in method deleteById,ids is null! ");
+		if (c == null || ids==null) return 0;
 		String sql = getObjToSQLRich().toDeleteByIdSQL(c, ids);
 		Logger.logSQL("deleteById SQL: ", sql);
 		_regEntityClass2(c);
@@ -346,7 +457,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 	@Deprecated
 	public <T> List<T> select(T entity, IncludeType includeType, Condition condition) {
 		if (entity == null) return null;
-		String sql = getObjToSQLRich().toSelectSQL(entity, includeType,condition);
+		String sql = getObjToSQLRich().toSelectSQL(entity, includeType, condition);
 		Logger.logSQL("select SQL: ", sql);
 		return getBeeSql().select(sql, entity);
 	}
@@ -357,8 +468,18 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		
 		String sql = getObjToSQLRich().toSelectSQL(entity, includeType,condition);
 		Logger.logSQL("selectJson SQL: ", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		
+		return getBeeSql().selectJson(sql);
+	}
+	
+	@Override
+	public <T> String selectJson(T entity, Condition condition) {
+		if (entity == null) return null;
+		
+		String sql = getObjToSQLRich().toSelectSQL(entity, condition.getIncludeType(),condition);
+		Logger.logSQL("selectJson SQL: ", sql);
+		_regEntityClass1(entity);
 		return getBeeSql().selectJson(sql);
 	}
 
@@ -368,7 +489,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		int r = 0;
 		String sql = getObjToSQLRich().toUpdateBySQL(entity, whereFields); //updateBy
 		Logger.logSQL("update SQL(whereFields) :", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		r = getBeeSql().modify(sql);
 
 		return r;
@@ -380,7 +501,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		int r = 0;
 		String sql = getObjToSQLRich().toUpdateBySQL(entity, whereFields, includeType);//updateBy
 		Logger.logSQL("update SQL(whereFields) :", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		r = getBeeSql().modify(sql);
 
 		return r;
@@ -393,10 +514,16 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		int r = 0;
 		String sql = getObjToSQLRich().toUpdateBySQL(entity, whereFields, condition);//updateBy
 		Logger.logSQL("update SQL(whereFields) :", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		r = getBeeSql().modify(sql);
 
 		return r;
+	}
+	
+	//v1.9
+	@Override
+	public <T> int updateById(T entity, Condition condition) {
+       return updateBy(entity, "id", condition);
 	}
 
 	//v1.7.2
@@ -406,7 +533,7 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		int r = 0;
 		String sql = getObjToSQLRich().toUpdateSQL(entity, updateFields, condition);
 		Logger.logSQL("update SQL(updateFields) :", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		r = getBeeSql().modify(sql);
 		
 		return r;
@@ -419,19 +546,126 @@ public class ObjSQLRich extends ObjSQL implements SuidRich {
 		int r = 0;
 		String sql = getObjToSQLRich().toUpdateSQL(entity, "", condition);
 		Logger.logSQL("update SQL(condition) :", sql);
-		_regEntityClass(entity);
+		_regEntityClass1(entity);
 		r = getBeeSql().modify(sql);
 		
 		return r;
 	}
 	
-	private <T> void _regEntityClass(T entity){
+	//1.9
+	@Override
+	public <T> int insert(List<T> entityList) {
+		if (entityList == null || entityList.size()<1) return -1;
+		T entity[]=toEntityArray(entityList);
+		return insert(entity);
+	}
+	
+	//1.9
+	@Override
+	public <T> int insert(List<T> entityList, int batchSize) {
+		if (entityList == null || entityList.size()<1) return -1;
+		T entity[]=toEntityArray(entityList);
+		return insert(entity, batchSize);
+	}
+
+	//1.9
+	@Override
+	public <T> int insert(List<T> entityList, String excludeFields) {
+		if (entityList == null || entityList.size()<1) return -1;
+		T entity[]=toEntityArray(entityList);
+		return insert(entity, excludeFields);
+	}
+
+	//1.9
+	@Override
+	public <T> int insert(List<T> entityList, int batchSize, String excludeFields) {
+		if (entityList == null || entityList.size()<1) return -1;
+		T entity[]=toEntityArray(entityList);
+		
+		return insert(entity, batchSize, excludeFields);
+	}
+
+	//没能将entity传到SqlLib,需要注册
+	private <T> void _regEntityClass1(T entity){
 		if(entity==null) return ;
 		HoneyContext.regEntityClass(entity.getClass());
 	}
 	
+	@SuppressWarnings("rawtypes")
 	private void _regEntityClass2(Class clazz){
 		HoneyContext.regEntityClass(clazz);
+	}
+	
+	@SuppressWarnings("unchecked")
+	private <T> T[] toEntityArray(List<T> entityList) {
+		
+		int len=entityList.size();
+		T entity[]=(T[])new Object[len];
+		
+		for (int i=0; i < len; i++) {
+			entity[i]=entityList.get(i);
+		}
+		
+		return entity;
+	}
+	
+	@Override
+	public SuidRich setDynamicParameter(String para, String value) {
+		OneTimeParameter.setAttribute(para, value);
+		return this;
+	}
+	
+	@Override
+	public <T> boolean exist(T entity) {
+		int r = count(entity);
+		return r > 0 ? true : false;
+	}
+
+	@Override
+	public <T> int update(T oldEntity, T newEntity) {
+
+		if (oldEntity == null || newEntity == null) return -1;
+		String oldEntityFullName = oldEntity.getClass().getName();
+		String newEntityFullName = newEntity.getClass().getName();
+		if (!oldEntityFullName.equals(newEntityFullName)) {
+			throw new BeeErrorGrammarException(
+					"BeeErrorGrammarException: the oldEntity and newEntity must be same type!");
+		}
+
+		Map<String, Object> oldMap = SuidHelper.entityToMap(oldEntity);
+		Map<String, Object> newMap = SuidHelper.entityToMap(newEntity);
+
+		MapSql updateMapSql = BeeFactoryHelper.getMapSql();
+		updateMapSql.put(MapSqlKey.Table, _toTableName(oldEntity));
+		updateMapSql.put(MapSqlSetting.IsNamingTransfer, true);
+		updateMapSql.put(oldMap);
+		updateMapSql.putNew(newMap);
+
+		Logger.logSQL("update(T oldEntity, T newEntity) with MapSuid, ", "");
+		MapSuid mapSuid = BeeFactoryHelper.getMapSuid();
+		return mapSuid.update(updateMapSql);
+	}
+	
+	private static String _toTableName(Object entity) {
+		return NameTranslateHandle.toTableName(NameUtil.getClassFullName(entity));
+	}
+	
+	/**
+	 * 保存一个实体(一条记录).
+	 * 如果可以区分开,建议明确调用insert(entity)或者update(entity),这样更加安全和高效.
+	 * @param entity
+	 * @return 返回受影响的行数.
+	 * @since 1.9.8
+	 */
+	@Override
+	public <T> int save(T entity) {
+		Object id = HoneyUtil.getIdValue(entity);
+		if (id == null) return insert(entity);
+		Object one = selectById(entity, id.toString());
+		if (one != null)
+			return update(entity);
+		else
+			return insert(entity);
 	}
 
 }
